+++ AndroidManifest.xml
@@ -252,5 +252,9 @@
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
         </activity>
+        <meta-data
+            android:name=
+                "com.google.android.gms.cast.framework.OPTIONS_PROVIDER_CLASS_NAME"
+            android:value="com.firebirdberlin.nightdream.CastOptionsProvider" />
     </application>
 </manifest>
