+++ build.gradle
@@ -42,6 +42,8 @@ dependencies {
     implementation 'androidx.exifinterface:exifinterface:1.3.2'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
     implementation "androidx.coordinatorlayout:coordinatorlayout:1.1.0"
+    implementation 'androidx.mediarouter:mediarouter:1.2.2'
+    implementation 'com.google.android.gms:play-services-cast-framework:19.0.0'
     testImplementation 'junit:junit:4.13.2'
 }
 
@@ -60,6 +62,11 @@ def getApiKeyDarkSky() {
 android {
     compileSdkVersion 30
     buildToolsVersion '29.0.2'
+
+    compileOptions {
+        sourceCompatibility 1.8
+        targetCompatibility 1.8
+    }
 
     sourceSets {
         main {
