+++ res/drawable/webradiopanelborder.xml
@@ -5,7 +5,7 @@
     android:right="0dp"
     android:bottom="-1dp">
 
-    <shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+    <shape android:shape="rectangle">
         <corners
             android:topLeftRadius="30dp"
             android:topRightRadius="30dp" />
