+++ res/layout-v17/main.xml
@@ -105,7 +105,7 @@
             </LinearLayout>
         </LinearLayout>
 
-        <FrameLayout
+        <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="0dp"
             android:layout_weight="1">
@@ -130,7 +130,7 @@
             </com.firebirdberlin.nightdream.ui.ClockLayoutContainer>
 
             <include layout="@layout/side_panel" />
-        </FrameLayout>
+         </androidx.constraintlayout.widget.ConstraintLayout>
 
         <com.firebirdberlin.nightdream.ui.BottomPanelLayout
             android:id="@+id/bottomPanel"
