+++ src/com/firebirdberlin/nightdream/CastOptionsProvider.java
@@ -1,0 +1,25 @@
+package com.firebirdberlin.nightdream;
+
+import android.content.Context;
+
+import com.google.android.gms.cast.CastMediaControlIntent;
+import com.google.android.gms.cast.framework.CastOptions;
+import com.google.android.gms.cast.framework.OptionsProvider;
+import com.google.android.gms.cast.framework.SessionProvider;
+
+import java.util.List;
+
+public class CastOptionsProvider implements OptionsProvider {
+    @Override
+    public CastOptions getCastOptions(Context context) {
+        CastOptions castOptions = new CastOptions.Builder()
+                .setReceiverApplicationId(CastMediaControlIntent.DEFAULT_MEDIA_RECEIVER_APPLICATION_ID)
+                .build();
+        return castOptions;
+    }
+    @Override
+    public List<SessionProvider> getAdditionalSessionProviders(Context context) {
+        return null;
+    }
+}
+
